Move registration, don't set VAD
